feat: add filter + pagination

+++ services/listingService.js
@@ -6,8 +6,19 @@ export const createListing = async (data) => {
     return listing;
 }
 
-export const fetchAllListings = async () => {
-    const listings = await listingsModel.find();
+export const fetchAllListings = async (page, limit, location, price) => {
+    const filter = {};
+    page = +page;
+    limit = +limit;
+
+    if (location) {
+        filter.location = location;
+    }
+    if (price) {
+        filter.price = price;
+    }
+
+    const listings = await listingsModel.find(filter).skip((page - 1) * limit).limit(limit);
 
     return listings;
 }
@@ -35,7 +46,12 @@ export const updateListingInDB = async (data, id) => {
     return listing;
 }
 
-export const deleteListingInDB = async (id) => {
+export const deleteListingInDB = async (id, userId) => {
+    const isOwner = await listingsModel.findOne({ _id: id, host: userId })
+    if (!isOwner) {
+        throw new Error("You are not authorized!");
+    }
+
     const listing = await listingsModel.findByIdAndDelete(id);
 
     if (!listing) {
